243: move FixedSizeData
